refactored previous changes for SorterConfig (now called SorterData)
